event mode via config
