+++ core/src/main/java/ee/joonasvali/spaceblaster/core/Config.java
@@ -1,0 +1,8 @@
+package ee.joonasvali.spaceblaster.core;
+
+/**
+ * Contains config from spaceblaster.yaml
+ */
+public class Config {
+  public boolean eventMode;
+}
