+++ core/src/main/java/module-info.java
@@ -3,5 +3,6 @@ module ee.joonasvali.spaceblaster.core {
   requires gdx.freetype;
   requires slf4j.api;
   requires ee.joonasvali.spaceblaster.event;
+  requires org.yaml.snakeyaml;
   exports ee.joonasvali.spaceblaster.core;
 }
