+++ desktop/assembly/assembly-unix.xml
@@ -47,6 +47,17 @@
     </fileSet>
 
     <fileSet>
+      <directory>${basedir}/configuration</directory>
+      <filtered>true</filtered> <!-- Causes maven placeholders to be replaced in the files -->
+      <outputDirectory>/</outputDirectory>
+      <includes>
+        <include>**/*</include>
+      </includes>
+      <lineEnding>unix</lineEnding>
+      <fileMode>0755</fileMode>
+    </fileSet>
+
+    <fileSet>
       <directory>${basedir}/backgrounds</directory>
       <outputDirectory>/backgrounds</outputDirectory>
       <includes>
