+++ desktop/configuration/spaceblaster.yaml
@@ -1,0 +1,1 @@
+eventMode: false
